Reorganized reversing LinkedLists to be consistant
